Double.MAX_VALUE -> Float.MAX_VALUE everywhere As a comment suggested Double.MAX_VALUE can cause arithmetic problems

+++ flow/enginesrc/org/labkey/flow/analysis/model/FlowJo10_0_6Workspace.java
@@ -118,8 +118,8 @@ public class FlowJo10_0_6Workspace extends PC75Workspace
         for (RectangleGateDimensionType dim : xRectangleGate.getDimensionArray())
         {
             axes.add(dim.getFcsDimension().getName());
-            lstMin.add(dim.isSetMin() ? dim.getMin() : -Double.MAX_VALUE); // XXX: Comment in IntervalGate implies we use Float.MIN/MAX_VALUE instead
-            lstMax.add(dim.isSetMax() ? dim.getMax() : Double.MAX_VALUE);
+            lstMin.add(dim.isSetMin() ? dim.getMin() : -Float.MAX_VALUE);
+            lstMax.add(dim.isSetMax() ? dim.getMax() : Float.MAX_VALUE);
         }
 
         if (axes.size() == 1)
